move modal context to context and useModal to hooks
