+++ src/hooks/useModal.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import ModalContext from './ModalContext';
+import ModalContext from '../context/ModalContext';
 
 const useModal = () => {
     const context = useContext(ModalContext);
